Wire fully-quantized decode engine (308 tok/s) + fix 2 deploy bugs
The fast decode engine (StaticDecodeEngine) existed but was unreachable from the
default path: stp_bench runs the fp32 eager O(T^2) path (15 tok/s), the
deploy-quant build was never composed for trained checkpoints, and the engine's
bf16 deploy path had bit-rot. Now: 308 tok/s B=1, token-match 256/256
byte-identical to the eager-bf16 reference, 627 MB resident (20x over fp32 eager).

- packed_ternary_infer.to_deploy_inference(model): composes the deploy-quant build
  for an already-carved/QAT trained ckpt (int6 embeds + PackedTernary backbone
  projs + Int8RowLinear attention + 2-bit-packed carved MLPs) → deploy_quant=True
  so StaticDecodeEngine uses the fused int8/2bit/bf16 kernels. Run under bf16 autocast.
- materialize_top_level: pack TernarySTE-parametrized backbone projs via
  from_parametrized (reads the shadow) not from_linear (which re-derived the scale
  from the STE output — rel 0.33 wrong on lm_mixer; bit-exact now).
- kv_cache_static._emit: write the HCA compressed block in the destination buffer
  dtype (bf16 deploy) — was fp32 into a bf16 buffer (RuntimeError on deploy stack).
- fused_decode_step._front_gemm_kernel: pick the largest BK dividing KCH=KDIM/KS.
  The HAS_SC K-loop is unmasked; under deploy_quant (KS=8) the 276M KDIM=768 gives
  KCH=96 and the old BK=64 over-read past the k-slice -> corrupt int8 attention
  (engine diverged from eager at token 222). 30B path (KCH=1024) keeps BK=256;
  fp32 path (BK=64) byte-identical.

Gate: ignore/bench_decode.py -> DECODE_BENCH_PASS tok/s=308 match=256/256 (256 tok,
exercises CSA-pool + HCA emits) on tst_stp_off_50k/step_50000.pt, RTX 5090.

+++ morph/model/kv_cache_static.py
@@ -690,7 +690,7 @@ class StaticDecodeEngine:
         rows = torch.cat([(self.pos_dev - (m - 1) + self._ar127) % (self.W_xh - 1),
                           self._stage_row_hca])
         blk = impl.comp_norm(impl.compressor(s.X.index_select(1, rows)))[:, -1:]
-        s.C_comp.index_copy_(1, idx, blk)
+        s.C_comp.index_copy_(1, idx, blk.to(s.C_comp.dtype))
 
     def _attn_site(self, s: _Site, emit: bool) -> Tensor:
         """Site attention. The norm_attn output (x_t) is ALREADY in the X staging slot."""
